+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/main/MainActivity.kt
@@ -18,8 +18,8 @@ import com.woowacourse.staccato.databinding.ActivityMainBinding
 import com.woowacourse.staccato.presentation.base.BindingActivity
 import com.woowacourse.staccato.presentation.memory.MemoryFragment.Companion.MEMORY_ID_KEY
 import com.woowacourse.staccato.presentation.memorycreation.MemoryCreationActivity
+import com.woowacourse.staccato.presentation.moment.VisitFragment.Companion.VISIT_ID_KEY
 import com.woowacourse.staccato.presentation.util.showToast
-import com.woowacourse.staccato.presentation.visit.VisitFragment.Companion.VISIT_ID_KEY
 import com.woowacourse.staccato.presentation.visitcreation.VisitCreationActivity
 import com.woowacourse.staccato.presentation.visitcreation.VisitCreationActivity.Companion.MEMORY_TITLE_KEY
