+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/memory/MemoryFragment.kt
@@ -22,8 +22,8 @@ import com.woowacourse.staccato.presentation.memory.adapter.VisitsAdapter
 import com.woowacourse.staccato.presentation.memory.viewmodel.MemoryViewModel
 import com.woowacourse.staccato.presentation.memory.viewmodel.MemoryViewModelFactory
 import com.woowacourse.staccato.presentation.memoryupdate.MemoryUpdateActivity
+import com.woowacourse.staccato.presentation.moment.VisitFragment.Companion.VISIT_ID_KEY
 import com.woowacourse.staccato.presentation.util.showToast
-import com.woowacourse.staccato.presentation.visit.VisitFragment.Companion.VISIT_ID_KEY
 import com.woowacourse.staccato.presentation.visitcreation.VisitCreationActivity
 
 class MemoryFragment :
