+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/visitcreation/VisitCreationActivity.kt
@@ -13,8 +13,8 @@ import com.woowacourse.staccato.databinding.ActivityVisitCreationBinding
 import com.woowacourse.staccato.presentation.base.BindingActivity
 import com.woowacourse.staccato.presentation.common.PhotoAttachFragment
 import com.woowacourse.staccato.presentation.memory.MemoryFragment.Companion.MEMORY_ID_KEY
+import com.woowacourse.staccato.presentation.moment.VisitFragment.Companion.VISIT_ID_KEY
 import com.woowacourse.staccato.presentation.util.showToast
-import com.woowacourse.staccato.presentation.visit.VisitFragment.Companion.VISIT_ID_KEY
 import com.woowacourse.staccato.presentation.visitcreation.adapter.PhotoAttachAdapter
 import com.woowacourse.staccato.presentation.visitcreation.viewmodel.VisitCreationViewModel
 import com.woowacourse.staccato.presentation.visitcreation.viewmodel.VisitCreationViewModelFactory
