+++ android/Staccato_AN/app/src/main/res/layout/item_others_visit_log.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="visitLog"
-            type="com.woowacourse.staccato.presentation.visit.model.VisitDetailUiModel.VisitLogUiModel" />
+            type="com.woowacourse.staccato.presentation.moment.model.VisitDetailUiModel.VisitLogUiModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
