+++ android/Staccato_AN/app/src/main/res/layout/item_visit_default.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="visitDefault"
-            type="com.woowacourse.staccato.presentation.visit.model.VisitDetailUiModel.VisitDefaultUiModel" />
+            type="com.woowacourse.staccato.presentation.moment.model.VisitDetailUiModel.VisitDefaultUiModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
