+++ android/Staccato_AN/app/src/main/res/navigation/bottom_navigation_graph.xml
@@ -31,7 +31,7 @@
 
     <fragment
         android:id="@+id/visitFragment"
-        android:name="com.woowacourse.staccato.presentation.visit.VisitFragment"
+        android:name="com.woowacourse.staccato.presentation.moment.VisitFragment"
         android:label="fragment_visit"
         tools:layout="@layout/fragment_visit" />
